Restructured folders to represent workflows for easier sharing and maintainability.
This commit restructures the project folders to better represent the workflow and facilitate easier sharing of examples with prospects.
Changes include:
Introduced a top-level workflows folder to house separate subfolders for each workflow (e.g., ecommerce, fieldservices, plm).
Created actions and blueprints folders within each workflow folder to group related files.
Added a shared folder to hold common components like externalConstraints.
Updated import paths to reflect the new folder structure.
This restructuring improves organization, clarity, and maintainability of the codebase, making it easier for prospects to find relevant examples and for developers to work with the code.
