+++ src/shared/externalContraints/externalConstraints.ts
