+++ src/workflows/fieldServices/blueprints/inventory.ts
