+++ src/workflows/plm/blueprints/attributes.ts
